Move Worktree and directory functions to separate files
