+++ backend/infrahub/git/__init__.py
@@ -1,7 +1,5 @@
+from infrahub.git.directory import initialize_repositories_directory
 from infrahub.git.repository import (
-    BRANCHES_DIRECTORY_NAME,
-    COMMITS_DIRECTORY_NAME,
-    TEMPORARY_DIRECTORY_NAME,
     ArtifactGenerateResult,
     CheckDefinitionInformation,
     GraphQLQueryInformation,
@@ -9,15 +7,10 @@ from infrahub.git.repository import (
     InfrahubRepository,
     RepoFileInformation,
     TransformPythonInformation,
-    Worktree,
     extract_repo_file_information,
-    initialize_repositories_directory,
 )
 
 __all__ = [
-    "BRANCHES_DIRECTORY_NAME",
-    "COMMITS_DIRECTORY_NAME",
-    "TEMPORARY_DIRECTORY_NAME",
     "ArtifactGenerateResult",
     "InfrahubReadOnlyRepository",
     "InfrahubRepository",
@@ -25,7 +18,6 @@ __all__ = [
     "CheckDefinitionInformation",
     "RepoFileInformation",
     "GraphQLQueryInformation",
-    "Worktree",
     "extract_repo_file_information",
     "initialize_repositories_directory",
 ]
