+++ backend/infrahub/git/constants.py
@@ -1,0 +1,3 @@
+COMMITS_DIRECTORY_NAME = "commits"
+BRANCHES_DIRECTORY_NAME = "branches"
+TEMPORARY_DIRECTORY_NAME = "temp"
