+++ backend/tests/unit/git/test_git_repository.py
@@ -15,17 +15,15 @@ from infrahub.exceptions import (
     TransformError,
 )
 from infrahub.git import (
-    BRANCHES_DIRECTORY_NAME,
-    COMMITS_DIRECTORY_NAME,
-    TEMPORARY_DIRECTORY_NAME,
     ArtifactGenerateResult,
     CheckDefinitionInformation,
     GraphQLQueryInformation,
     InfrahubRepository,
     RepoFileInformation,
-    Worktree,
     extract_repo_file_information,
 )
+from infrahub.git.constants import BRANCHES_DIRECTORY_NAME, COMMITS_DIRECTORY_NAME, TEMPORARY_DIRECTORY_NAME
+from infrahub.git.worktree import Worktree
 from infrahub.utils import find_first_file_in_directory
 from tests.helpers.test_client import dummy_async_request
